requirements.txt
langchain==1.2.13
langchain-community==0.4.1
python-dotenv==1.2.2
chromadb==1.5.5
pypdf==6.9.2
pdf2image==1.16.3
pydantic==2.5.0
requests==2.31.0